WIP split off the BitwallModel into its own Require module; still needs debugging
